+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -1,17 +1,14 @@
 package com.techcourse;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
+import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapterImpl;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
-import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -19,7 +16,7 @@ public class DispatcherServlet extends HttpServlet {
 
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
-
+    HandlerAdapterImpl handlerAdapter = new HandlerAdapterImpl();
     private List<HandlerMapping> mappings;
 
     public DispatcherServlet() {
@@ -48,30 +45,11 @@ public class DispatcherServlet extends HttpServlet {
         try {
             for (HandlerMapping handlerMapping : mappings) {
                 final var handler = handlerMapping.getHandler(request);
-                if (handler != null) {
-                    if (handler instanceof HandlerExecution) {
-                        final var modelAndView = ((HandlerExecution) handler).handle(request, response);
-                        modelAndView.getView().render(modelAndView.getModel(), request, response);
-                    } else if (handler instanceof Controller) {
-                        final var viewName = ((Controller) handler).execute(request, response);
-                        move(viewName, request, response);
-                    } else {
-                        response.setStatus(HttpServletResponse.SC_NOT_FOUND);
-                    }
-                }
+                handlerAdapter.handle(request, response, handler);
             }
         } catch (Throwable e) {
             log.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
         }
-    }
-
-    private void move(
-            final String viewName,
-            final HttpServletRequest request,
-            final HttpServletResponse response
-    ) throws Exception {
-        final var view = new JspView(viewName);
-        view.render(Map.of(), request, response);
     }
 }
