+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerAdapter.java
@@ -1,0 +1,9 @@
+package com.interface21.webmvc.servlet.mvc.tobe;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+public interface HandlerAdapter {
+
+    void handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception;
+}
